feat: Refactor module and context handling with new redirect method and improved module initialization
- Added `redirect` method to the `Context` class for handling HTTP redirects.
- Refactored `Application` class to delegate module initialization to the `Injector` class.
- Moved module provider registration logic into the `Injector` class for better separation of concerns.
- Improved module initialization flow in the `Injector` class, including recursive module dependency resolution and controller instantiation.
- Updated `Injector` to handle controller routes and dependencies more efficiently.
- Removed redundant code from `Application` class regarding module provider registration.
- Cleaned up the `Module` decorator to simplify dependency injection logic.
